displayer_service/requirements.txt
astroid==2.15.8
autopep8==2.0.4
cfgv==3.3.1
dill==0.3.7
distlib==0.3.8
filelock==3.12.2
identify==2.5.24
importlib-metadata==6.7.0
inky==1.5.0
isort==5.11.5
lazy-object-proxy==1.9.0
mccabe==0.7.0
nodeenv==1.8.0
numpy==1.16.2
Pillow==9.5.0
platformdirs==4.0.0
pre-commit==2.21.0
pycodestyle==2.10.0
pylint==2.17.7
PyYAML==6.0.1
RPi.GPIO==0.7.1
smbus2==0.4.3
spidev==3.6
tomli==2.0.1
tomlkit==0.12.3
typed-ast==1.5.5
typing_extensions==4.7.1
virtualenv==20.25.0
wrapt==1.16.0
zipp==3.15.0
